Average Time Implementation

+++ manager/RequestManager.py
@@ -229,7 +229,7 @@ def buildArchive(uuid, renderRequest, metadata):
     renderArchive.project_name = metadata[0]
     renderArchive.hardware_stats = HardwareStats.from_dict(eval(metadata[1]))
     renderArchive.finish_time = metadata[2]
-    renderArchive.avg_frame = int(metadata[3])
+    renderArchive.avg_frame = float(metadata[3])
     renderArchive.frame_map = metadata[4].strip('][').split(', ')
     print(metadata[4])
     renderArchive.render_settings = RenderSettings.from_dict(eval(metadata[5]))
